requirements.txt
fastapi
uvicorn
requests
python-dotenv
email-validator